requirements.txt
vosk==0.3.45
sounddevice==0.4.6
numpy==1.26.4
pyttsx3==2.90
paho-mqtt==2.1.0
opencv-python==4.10.0.84
